Adding Notification application.
This application introduces how to use the NotificationManager and
NotificationCompat.Builder to create notifications. The user can
create a notification manually, or make it a repeating notification.

To make a repeating notification, the application also introduce
how to use the AlarmManager class along with an implementation of
a BroadcastReceiver that will be called when the alarm goes off.